Extracted RangeMaker into its own compilation unit.
