+++ src/appointment_ranges/appointment_ranges.h
@@ -5,11 +5,6 @@
 #include "range.h"
 #include <vector>
 
-const int c_lastMinute = 24 * 60;
-
 std::vector<Range> getRangesFromAppointments(const std::vector<Appointment>& appointments);
-
-Pattern getPatternFromState(FreeBusy state);
-double positionOfMinute(int minute);
 
 #endif /* end of include guard: APPOINTMENT_RANGES_INCLUDED */
